+++ org.eclipse.xtext.extras.tests/src/org/eclipse/xtext/resource/uriHell/FileInSourceAndPackagedProjectTest.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2013, 2020 itemis AG (http://www.itemis.eu) and others.
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.xtext.resource.uriHell;
+
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.xtext.resource.uriHell.AbstractURIHandlerWithEcoreTest;
+
+/**
+ * @author Sebastian Zarnekow - Initial contribution and API
+ */
+public class FileInSourceAndPackagedProjectTest extends AbstractURIHandlerWithEcoreTest {
+	@Override
+	public URI getResourceURI() {
+		return URI.createURI("platform:/resource/projectName/src/org/packageName/First.ecore");
+	}
+
+	@Override
+	public URI getPackagedResourceURI() {
+		return URI.createURI("platform:/resource/projectName/org/packageName/First.ecore");
+	}
+
+	@Override
+	public URI getReferencedURI() {
+		return URI.createURI("platform:/resource/other/org/packageName/Second.ecore");
+	}
+
+	@Override
+	public URI getPackagedReferencedURI() {
+		return URI.createURI("platform:/resource/other/org/packageName/Second.ecore");
+	}
+}
