+++ org.eclipse.xtext.extras.tests/src/org/eclipse/xtext/resource/uriHell/FilesInModelAndSourceMavenDifferentProjectsTest.java
