+++ org.eclipse.xtext.extras.tests/src/org/eclipse/xtext/resource/uriHell/FilesInSameSourceFolderTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2013, 2016 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2013, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -14,11 +14,10 @@ import org.eclipse.xtext.resource.uriHell.AbstractURIHandlerWithEcoreTest;
 /**
  * @author Sebastian Zarnekow - Initial contribution and API
  */
-@SuppressWarnings("all")
-public class FilesInSourceMavenAndModelDifferentProjectsTest extends AbstractURIHandlerWithEcoreTest {
+public class FilesInSameSourceFolderTest extends AbstractURIHandlerWithEcoreTest {
   @Override
   public URI getResourceURI() {
-    return URI.createURI("platform:/resource/projectName/src/main/org/package/First.ecore");
+    return URI.createURI("platform:/resource/projectName/src/org/package/First.ecore");
   }
   
   @Override
@@ -28,11 +27,11 @@ public class FilesInSourceMavenAndModelDifferentProjectsTest extends AbstractURI
   
   @Override
   public URI getReferencedURI() {
-    return URI.createURI("platform:/resource/other/model/Second.ecore");
+    return URI.createURI("platform:/resource/projectName/src/org/other/Second.ecore");
   }
   
   @Override
   public URI getPackagedReferencedURI() {
-    return URI.createURI("platform:/resource/other/model/Second.ecore");
+    return URI.createURI("platform:/resource/projectName/org/other/Second.ecore");
   }
 }
