+++ org.eclipse.xtext.extras.tests/src/org/eclipse/xtext/resource/uriHell/LoadedByClasspathBothFromMavenStructureTest2.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2013, 2016 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2013, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -14,15 +14,15 @@ import org.eclipse.xtext.resource.uriHell.LoadedByClasspathBothFromMavenStructur
 /**
  * @author Sebastian Zarnekow - Initial contribution and API
  */
-@SuppressWarnings("all")
 public class LoadedByClasspathBothFromMavenStructureTest2 extends LoadedByClasspathBothFromMavenStructureTest {
-  @Override
-  public URI getResourceURI() {
-    return URI.createURI("platform:/resource/org.eclipse.xtext.tests/src/main/org/eclipse/xtext/resource/ecore/utf8.ecore");
-  }
-  
-  @Override
-  public URI getPackagedResourceURI() {
-    return URI.createURI("classpath:/org/eclipse/xtext/resource/ecore/utf8.ecore");
-  }
+	@Override
+	public URI getResourceURI() {
+		return URI.createURI(
+				"platform:/resource/org.eclipse.xtext.tests/src/main/org/eclipse/xtext/resource/ecore/utf8.ecore");
+	}
+
+	@Override
+	public URI getPackagedResourceURI() {
+		return URI.createURI("classpath:/org/eclipse/xtext/resource/ecore/utf8.ecore");
+	}
 }
